Main svc

+++ stats/stats-server/src/main/java/ru/practicum/service/StatsService.java
@@ -50,10 +50,12 @@ public class StatsService {
             var item = new HitOutput();
             item.setApp("ewm-main-service");
             item.setUri(pair.getKey());
-            item.setHits((long) hits.size());
+            item.setHits(hits.size());
             res.add(item);
         }
 
         return res.stream().sorted(Comparator.comparingLong(hitOutput -> -hitOutput.getHits())).collect(Collectors.toList());
     }
+
+
 }
